chore: move SubsectionEntry to molecule
